refactor(llma): extract PlaygroundSaveMenu into its own file
Move the save menu component out of the scene file into
PlaygroundSaveMenu.tsx. Replace the nested ternary for
linkedContextLabel with a clearer helper function.
